THE FRAMEWORK FOR A STATIC WEBSITE
These folders are the basic HTML that creates the framework for the deployment of the information onto the static website. It is critical that these are not adjusted without comparing them to the .css folder in the RESULTS folder as the .css folder styles the static elements of the website.