THE SOURCE FILES FOR THE WEBSITE
These files are the source files that the deployment codes points to when building the website for deployment purposes. Any files placed in this folder can be written to the website as pages/ links in the header information. This is not automated yet, until it is automated then the HTML NAVIGATION folder will need to be updated with the correct .md file list.