POINTER FILE FOR THE HTML FILES AFTER DEPLOYMENT
This is where the converted files will be sent to after the DEPLOYMENT.PY file is run. It contains a .CSS file - DO NOT ALTER THIS FILE. The .css file is critical for the styling of the static website. Any other content in this folder will change as the source folder changes and the site is built.