IMAGES FOR THE .md FILES
This folder has the images that the .md files make reference to when building converting the files from .md to .html. Please store all images and EDA in this foler follwing after running the code. All images file names must be clear and concise to be used in the .md files when compiling the static website.